Isolated and deduplicated StoreObjs
