+++ src/reader.rs
@@ -12,6 +12,7 @@ use crate::error::{Error, Result};
 use crate::fileformat::{Blob, BlobHeader};
 use crate::objects::{OsmId, OsmObj};
 use crate::osmformat::PrimitiveBlock;
+use crate::store_objs::StoreObjs;
 use protobuf::Message;
 use pub_iterator_type::pub_iterator_type;
 use std::collections::btree_map::BTreeMap;
@@ -19,24 +20,6 @@ use std::collections::BTreeSet;
 use std::convert::From;
 use std::io::{self, Read};
 use std::iter;
-
-/// Trait to allow generic objects (not just BTreeMap) in some methods.
-pub trait StoreObjs {
-    /// Insert given object at given key index.
-    fn insert(&mut self, key: OsmId, value: OsmObj);
-    /// Check if object contains the given key.
-    fn contains_key(&self, key: &OsmId) -> bool;
-}
-
-impl StoreObjs for BTreeMap<OsmId, OsmObj> {
-    fn insert(&mut self, key: OsmId, value: OsmObj) {
-        self.insert(key, value);
-    }
-
-    fn contains_key(&self, key: &OsmId) -> bool {
-        self.contains_key(key)
-    }
-}
 
 /// The object to manage a pbf file.
 pub struct OsmPbfReader<R> {
